requirements.txt
asgiref==3.8.1
Django==4.1.1
djangorestframework==3.13.1
psycopg2==2.9.9
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
